ctrl: move common functions away
add a common source code file to hold functions
are used across all the controllers in this package.

Please note we should refrain to add pkgs name `common`,
but source code files is still OK.

Along the way, add `reasonFromError`, which allows us
to stop aliasing reason=condition,
which was a longstanding issue.
