+++ controllers/controllers.go
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2021 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package controllers
+
+import "errors"
+
+func reasonFromError(err error) string {
+	if err == nil {
+		return "AsExpected"
+	}
+	return "InternalError"
+}
+
+func messageFromError(err error) string {
+	if err == nil {
+		return ""
+	}
+	unwErr := errors.Unwrap(err)
+	if unwErr == nil {
+		return ""
+	}
+	return unwErr.Error()
+}
